New script to run benchmarks, seperate process for single workload

+++ run.py
@@ -29,7 +29,6 @@ sizes = [
     '1TB',
 ]
 
-# Worklods are filtred in case of `drop_caches` is true, otherwise runs all workloads
 workload_names = [
     'Init',
     'Read',
@@ -40,6 +39,8 @@ workload_names = [
     'ReadInsert_95_5',
     'Remove',
 ]
+
+destory_dbs = len(threads) > 1 or len(sizes) > 1
 
 
 def get_db_config_path(db_name, size):
@@ -56,28 +57,18 @@ def drop_system_caches():
     subprocess.run(['sudo', 'sh', '-c', '/usr/bin/echo', '3', '>', '/proc/sys/vm/drop_caches'])
 
 
-def run(db_name, size, threads_count, workload_name=None):
+def run(db_name, size, threads_count, workload_name):
     config_path = get_db_config_path(db_name, size)
     workloads_path = get_worklods_path(size)
-    process = None
 
-    if workload_name:
-        process = subprocess.Popen([
-                './build_release/bin/_ucsb_bench',
-                '-db', db_name,
-                '-c', config_path,
-                '-w', workloads_path,
-                '-threads', str(threads_count),
-                '-filter', workload_name,
-            ],
-            stdout=subprocess.PIPE)
-    else:
-        process = subprocess.Popen([
+    process = subprocess.Popen([
             './build_release/bin/_ucsb_bench',
             '-db', db_name,
             '-c', config_path,
             '-w', workloads_path,
-            '-threads', str(threads_count)],
+            '-threads', str(threads_count),
+            '-filter', workload_name,
+        ],
         stdout=subprocess.PIPE)
 
     # Print output
@@ -87,16 +78,14 @@ def run(db_name, size, threads_count, workload_name=None):
         if not line:
             break
 
-
 for threads_count in threads:
     for size in sizes:
+        if destory_dbs:
+            shutil.rmtree('./tmp/')
         for db_name in db_names:
-            if drop_caches:
-                for workload_name in workload_names:
+            for workload_name in workload_names:
+                if drop_caches:
                     drop_system_caches()
                     time.sleep(8)
-                    run(db_name, size, threads_count, workload_name)
-            else:
-                run(db_name, size, threads_count)
-        shutil.rmtree('./tmp/')
+                run(db_name, size, threads_count, workload_name)
